+++ package.json
@@ -17,13 +17,14 @@
     "@turf/random": "^5.1.5",
     "amqplib": "^0.5.1",
     "antd": "^2.13.8",
-    "async": "^2.0.0",
+    "async": "^2.6.0",
     "bcrypt-nodejs": "0.0.3",
+    "bluebird": "^3.5.1",
     "body-parser": "^1.15.2",
     "bootstrap-slider": "^9.7.0",
     "connect-flash": "~0.1.1",
     "cookie-parser": "~1.4.3",
-    "cookie-session": "^2.0.0-alpha.2",
+    "cookie-session": "^2.0.0-beta.3",
     "d3-queue": "^3.0.1",
     "dbgeo": "0.1.1",
     "debug": "~2.2.0",
