+++ public/javascripts/progressWidget.js
@@ -1,696 +1,6 @@
-function progressWidgetInit() {
-    $.widget("custom.progressWidget", {
-        options: {
-            value: 0
-        },
-
-        _create: function () {
-            /*
-            Initialize variables
-            */
-            this.element;
-            this.$el = $(this.element);
-            this.$anchor = $('.progress-tracker')
-            this.$flashhandler = $('#flashes');
-            var savedState = window.localStorage.getItem('progressWidget');
-            this.recentlyDeleted = [];
-            this.inLoop = false;
-            this.pollTimeout = null;
-            this.flashes = [];
-            this.flashTimeout = 5000;
-            this.pageReload = false;
-            this.uploadId = null;
-            this.ajaxInProgress = true;
-            this.reset = false;
-            /*
-            This loads savedState onto this.state; from window.localStorage
-            If you can't find it, just load defaultState.
-            */
-            if (savedState && savedState !== 'undefined') {
-                this.state = JSON.parse(savedState);
-                if (this.options.value && this.state.selectedIDs.indexOf(this.options.value) == -1) { //Ignore; this is test code.
-                    this.state.selectedIDs.shift(this.options.value);
-                }
-            }
-            else {
-                this.state = this._defaultState();
-            }
-            /*
-            We do a full refresh.
-            */
-            this.state.fullRefresh = 5;
-            /*
-            We check the URL to see if we have to query any ID's that have recently been uploaded.
-            TODO: change the nature of voxels/shapes dichotomy in pollquery. MAKE it so that the query goes /voxels/voxels?voxels=something& idk...
-            */
-            var urlString = window.location.pathname;
-            if (urlString.indexOf('/layers') != -1 || urlString.indexOf('/voxels') != -1) {
-                var isVoxel = (urlString.indexOf('/voxels') != -1);
-                this.state.pollType = isVoxel ? "voxels" : "shapes";
-                if (urlString.match(/\/\d+(?!\.)\/\d+(?!\.)/) != null) { // match two numbers with no decimal points
-                    try {
-                        var temp = urlString.split('/');
-                        this.uploadId = temp[temp.length - 1];
-                        if (!isVoxel){
-                            this.uploadId = parseInt(this.uploadId);
-                        }
-                        //IF IT IS A VOXEL, then "uploadId" is of the form id1$$id2$$etc. Remember that now.
-                        //if(typeof this.uploadId === 'number'){
-                            this.state.selectedIDs.unshift(this.uploadId);
-                        //}
-                    }
-                    catch (e) {
-                        console.log('PROGRESSWIDGET ERROR: error adding ID, url: ' + urlString)
-                        console.log(this.uploadId)
-                    }
-                }
-            }
-            /*
-            When we upload a file, pageReload becomes true, and the page will refresh after 1000ms.
-            */
-            if (this.uploadId != null) {
-                this.uploadId = null;
-                this.pageReload = true;
-            }
-
-            // } else {
-                this.refresh();
-            // }
-
-            /*
-            Bind the UI action
-            */
-            this._bindUIActions();
-            /*
-            If we display the progressWidget, we now have to refresh every 300ms.
-            */
-            if (this.state.isDisplayed) {
-                this.state.pollingState = 1;
-                this.$el.show()
-            }
-            else {
-                this.state.pollingState = 0;
-                this.$el.hide();
-            }
-            this._xhrLoop();
-        },
-
-        /*
-        Creates a flash.
-        */
-        _createFlash: function (input) {
-            this.$anchor.trigger('pflash', ["\t Progress Tracker: ", input]);
-        },
-
-        /*
-        Called from _refresh.
-        Self-explanatory.
-        */
-        _createWidgetMenu: function () {
-            /*
-            If we're not adding anything new, and we're just querying for currentJobs to add to the menu,
-            we partialRefresh.
-            */
-            if (this.state.fullRefresh != 5 && this.state.pollingState != 0 && !this.pageReload) {
-                this._partialRefresh();
-                return
-            }
-            var $progressWidget = this.element,
-                opts = this.options,
-                widgetCaption = "This tracks all current jobs in the worker queue. hover over the job names to see how many shapes/voxels have been processed.",
-                refreshCaption = "Click here to reset the Progress Tracker.",
-                htmlBuilder = '';
-
-
-            htmlBuilder += `<div id="widget-menu">
-                <span class="tooltiper caption-left" data-tooltip="${widgetCaption}"><i class="fa fa-question" style="top:'0px!important'"/></span>
-                <span class="tooltiper caption-right" data-tooltip="${refreshCaption}"><i class="fa fa-refresh refresh-widget" style="top:'0px!important'"/></span>
-                <ul id="jobList"></ul>
-            </div>
-            `
-            this.state.currentJobs;
-            widgetMenu = $.parseHTML(htmlBuilder);
-            widgetMenu = $(widgetMenu);
-            widgetMenu.text = widgetCaption;
-            /*
-            Adds the job display to the widgetMenu, if there are any.
-            */
-            if (this.state.currentJobs.length || this.state.completedJobs.length) {
-                if(this.state.currentJobs.length) {
-                    this.state.currentJobs.forEach(job => {
-                        htmlBuilder = createProgressBar(...job);
-                        var temp = $.parseHTML(htmlBuilder);
-                        widgetMenu.find('#jobList').append(temp);
-                    })
-                }
-                if(this.state.completedJobs.length) {
-                    this.state.completedJobs.forEach(job => {
-                        var id = job[0], name = job[1];
-                        htmlBuilder = createCompleteBar(id, name);
-                        var temp = $.parseHTML(htmlBuilder);
-                        widgetMenu.find('#jobList').append(temp);
-                    })
-                }
-            }
-            else {
-                widgetMenu.find('#jobList').append($.parseHTML(noJobs()));
-            }
-            /*
-            Now we remove the existing widgetMenu, and replace it with the widgetMenu we've just created.
-            */
-            if ($('#widget-menu').length)
-                $('#widget-menu').remove();
-
-            this.$widgetMenu = widgetMenu;
-            this.$el.append(this.$widgetMenu);
-            /*
-            Add event handlers for: closing out jobs, and resetting the progressTracker
-            */
-            if (this.$widgetMenu.children().length) {
-                $('.close-btn').on('click', this._closeItem.bind(this));
-            }
-
-            this._on($('.refresh-widget'), {
-                click: '_toggleReset'
-            })
-            /*
-            Now we only have to partialRefresh.
-            */
-            this.state.fullRefresh = 0;
-            toolTiper();
-        },
-
-        /*
-        This updates only the percentage in the progress bar.
-        Called from _createWidgetMenu.
-        */
-        _partialRefresh: function () {
-            this.state.currentJobs.forEach(job => {
-                var [id, jobName, numerator, denominator] = job,
-                    percentage = ((numerator / denominator) * 100).toFixed(2).toString(),
-                    innerString = jobName + ": " + numerator + "/" + denominator,
-                    frac = numerator + "/" + denominator,
-                    menuItem = $('#' + id),
-                    datatool = menuItem.find('.tooltiper').attr('data-tooltip', frac)
-                menuItem.find('.progress-label').text(percentage + '%');
-                menuItem.find('.progress-bar').css('width', percentage + '%');
-            });
-        },
-
-        /*
-        This removes the display of jobs. It tells us to stop querying for them.
-        */
-        _closeItem: function (e) {
-            var menuItem = $(e.target).closest('.menu-item');
-            var id = menuItem.attr('id')
-            menuItem.remove();
-
-            if (id === null) {
-                console.log('UPLOAD WIDGET Error: False destroychild id' + id)
-                this._createFlash('Error destroying child, reset local storage or contact developers if this persists.')
-                return false;
-            }
-
-            var inCurrentJobs = (this.state.selectedIDs.indexOf(id) !== -1);
-
-            if(inCurrentJobs){ // if its in the current jobs list
-                var delId = this.state.selectedIDs.splice(this.state.selectedIDs.indexOf(id), 1);
-                this.recentlyDeleted.push(delId);
-                // update removal for async calls
-                var delJob = this.state.currentJobs.filter((item) => {
-                    if (item[0] == id) {
-                        return item;
-                    }
-                });
-                var removed = this.state.currentJobs.splice(this.state.currentJobs.indexOf(delJob), 1);
-            }
-            else { // if its in the completed job list
-                this.recentlyDeleted.push(id);
-                // update removal for async calls
-                var delJob = this.state.completedJobs.filter((item) => {
-                    if (item[0] == id) {
-                        return item;
-                    }
-                });
-                var removed = this.state.completedJobs.splice(this.state.currentJobs.indexOf(delJob), 1);
-            }
-            this.refresh();
-            e.preventDefault();
-        },
-
-        /*
-        Called from _create.
-        This adds every event listener.
-        */
-        _bindUIActions: function () {
-            /*
-            Binds an event listener, so that the display of the widget can be toggled.
-            */
-            this._on(this.$anchor, {
-                click: '_toggleDisplay'
-            })
-            /*
-            This is the destructor, when we unload a page.
-            */
-            window.addEventListener('beforeunload', this._destroy.bind(this), false);
-            /*
-            This shows the flashes, whenever _createFlash is called.
-            */
-            this._on(this.$anchor, {
-                'pflash': function () {
-                    var event = arguments[1],
-                        message = arguments[2];
-
-                    var flash = $('<div class="pflash alert alert-info alert-dismissible"> <strong>' + event + '</strong>' + message + '</div>');
-
-                    flash.on('click', function () {
-                        $(this).fadeOut(1000, function() {
-                            $(this).remove();
-                        });
-                    });
-                    var t = setTimeout(function () {
-                        $(flash).fadeOut(1000, function () {
-                            $(this).remove();
-                        });
-                        this.flashes.shift();
-                    }.bind(this).bind(flash), this.flashTimeout);
-                    this.flashes.push(t);
-                    $('.navbar-fixed-top').append(flash);
-                }
-            });
-        },
-        /*
-        Unloads the widget: saves the state and removes all components.
-        */
-        _destroy: function () {
-            var test = JSON.stringify(this.state);
-            /*
-            Save the state to localStorage if this.reset is False, and vice versa.
-            */
-            if(!this.reset) {
-                window.localStorage.setItem('progressWidget', JSON.stringify(this.state));
-            }
-            else {
-                window.localStorage.removeItem('progressWidget');
-            }
-            var temp = window.localStorage.getItem('progressWidget');
-
-
-            if (this.state.isDisplayed)
-                this.$el.hide();
-
-            if (this.$widgetMenu) {
-                this.$widgetMenu.remove();
-            }
-            if (this.flashes.length) {
-                this.flashes.forEach(item => {
-                    clearTimeout(item);
-                })
-                // $(".pflash").each(e, i => {
-                //     $(i).remove();
-                // });
-            }
-            if (this.$el && this.$el.children().length)
-                this.$el.children().remove();
-        },
-
-        /*
-        Toggles this.reset, which influences if we want to save this.state to window.localStorage.
-        */
-        _toggleReset: function(e) {
-            if (e)
-                e.preventDefault();
-
-            this.reset = !this.reset;
-
-            if(this.reset)
-                this._createFlash("State will reset on page refresh. To cancel, click the refresh button again");
-            else
-                this._createFlash("State will no longer reset on page refresh.");
-            return
-        },
-        /*
-        Toggles display, which influences this.pollingState.
-        */
-        _toggleDisplay: function (e) {
-            if (e)
-                e.preventDefault();
-            /*
-            If you were displayed, don't display. Refresh only once every 5 seconds.
-            */
-            if (this.state.isDisplayed) {
-                this.state.pollingState = 0;// set polling var to false
-                this.$el.slideUp(200);
-                this.state.isDisplayed = !this.state.isDisplayed;
-            /*
-            If you weren't display, do display.
-            Refresh faster if selectedIDs is nontrivial (aka set pollingState = 1)
-            */
-            } else {
-                this.$el.slideDown(200);
-                this.state.isDisplayed = !this.state.isDisplayed;
-
-                if (this.state.selectedIDs.length !== 0) {
-                    if (!this.ajaxInProgress) {
-                        this.ajaxInProgress = true;
-                        this._xhrLoop();
-                    } else if (!this.inLoop && this.request === null) {
-                        clearTimeout(this.pollTimeout)
-                        this._xhrLoop();
-                    }
-                    this.state.pollingState = 1;
-                }
-                this.state.timeouts = 5;
-            }
-        },
-
-        //refresh state, should be triggered on init, f input (deletion), update jobs,=
-        refresh: function () {
-            /*
-            Update the widget DOM.
-            */
-            this._createWidgetMenu();
-            /*
-            If we have a new upload, or currentJobs is nontrivial, then we branch to else.
-            */
-            if (this.state.currentJobs.length === 0 && !this.pageReload) {
-                this.state.pollingState = 0;
-                this.state.selectedIDs = [];
-            }
-            else {
-                // Basically, take everything in currentJobs not in recentlyDeleted and add it to selectedIDs.
-                var temp = this.state.currentJobs.map(function (obj) {
-                    if (this.recentlyDeleted.indexOf(obj[0]) === -1)
-                        return obj[0];
-                }.bind(this));
-                if(temp.length)
-                    this.state.selectedIDs.concat(temp);
-
-                temp = new Set(this.state.selectedIDs);
-                this.state.selectedIDs = Array.from(temp);
-            }
-            /*
-            Set pollingState to be 1 if we're displayed and we have IDs to query the controller for.
-            */
-            if (this.state.selectedIDs.length === 0) {
-                this.state.pollingState = 0;
-            } else if (this.state.selectedIDs.length !== 0 && this.state.isDisplayed) {
-                this.state.pollingState = 1;
-            }
-            /*
-            Go straight to the xhrLoop if we don't have any requests, we don't have any ID's
-            and we're not already in a xhrLoop.
-            */
-            if(!this.request && !this.inLoop && !this.pageReload)
-                this._xhrLoop();
-        },
-
-        /*
-        Strictly called from _createXhrRequest.
-        This updates the jobs that we got from the controller.
-        */
-        _updateJobs: function (arr) {
-            var newJobs = [];
-            var delId, index, complete;
-
-
-            if (!arr)
-                return false; // don't update arr
-
-            for (var j = 0; j < arr.length; j++) { // can be length 1 if it is length 0 return
-                var input = arr[j];
-                if (input.length == 0)
-                    continue; //continue when no vals
-
-                if ( input.length === 1) { //error message
-                    input = input[0]; // unwrap string
-
-                    if(input.match(/\~\/\~/) != null ) {
-                        input = input.substr( input.indexOf('~/~') + 4, input.length)
-                        this._createFlash(input);
-                        continue;
-                    }
-
-                    var error = input.split('$$'),
-                        jobname;
-                    try {
-                        jobName = error[1].substring(error[1].lastIndexOf("/") + 1, error[1].length);
-                    } catch (e) {
-                        console.log(e)
-                        continue;
-                    }
-
-
-                    for (var i = 0; i < this.state.currentJobs.length; i++) {
-
-                        if (jobName.indexOf(this.state.currentJobs[i][1]) != -1) {
-                            this.state.fullRefresh = 5;
-                            complete = this.state.currentJobs.splice(i, 1);
-                            complete = complete[0];
-                            this.state.completedJobs.push(complete);
-                            index = this.state.selectedIDs.indexOf(complete[0])
-                            delId = this.state.selectedIDs.splice(index, 1); // stop querying this job
-                            this.recentlyDeleted.push(delId);
-
-                            this._createFlash(jobName + " has been removed from the queue");
-                            break;
-                        }
-                        else if(i == this.state.currentJobs.length -1 ){
-                            this._createFlash(jobName + " was not found in the queue. Error on processing initialization")
-                        }
-                    } // in case we want to make them stay in the queue until they're been x'ed out. (window.localStorage)
-                }
-
-                else if (input.length == 3) {
-                    var inArrs = false;
-                    for (var i = 0; i < this.state.currentJobs.length; i++) {
-                        if (input[0] == this.state.currentJobs[i][0]) {
-
-                            complete = this.state.currentJobs.splice(i, 1);
-                            //complete = complete[0];
-                            this.state.completedJobs.push(complete);
-                            index = this.state.selectedIDs.indexOf(complete[0])
-                            delId = this.state.selectedIDs.splice(index, 1); // stop querying this job
-                            this.recentlyDeleted.push(delId);
-                            this.state.fullRefresh = 5;
-
-                            if (input[2] === true) {
-                                var hasCompleted = false;
-                                this.state.completedJobs.map(job => {
-                                    if(job[0] === complete[0]){
-                                        hasCompleted = true;
-                                    }
-                                });
-
-                                if(!hasCompleted)
-                                    this.state.completedJobs.push(complete);
-
-                                this._createFlash(input[1] + " has completed!");
-                                this.complete = true;
-                                inArrs = true;
-                            }
-
-                            else {
-                                // for testing only
-                                // This will never show up in the code. What were you thinking?
-                                this._createFlash(input[1] + " has been removed from the Tracker - Data format was invalid!");
-                            }
-                            inArrs = true;
-                            break;
-
-                        }
-
-                        if(!inArrs && i === this.state.currentJobs.length -1 ) {
-                            index = this.state.selectedIDs.indexOf(input[0])
-                            delId = this.state.selectedIDs.splice(index, 1); // stop querying this job
-                            this.state.completedJobs.push(input);
-                            inArrs = true;
-                            this._createFlash(input[1] + " has completed!");
-                            this.complete = true;
-                        }
-                    }
-                    if(!inArrs && this.state.currentJobs.length === 0){ // for loop doesn't fire if its tempty
-                        index = this.state.selectedIDs.indexOf(input[0])
-                        delId = this.state.selectedIDs.splice(index, 1); // stop querying this job
-                        this.recentlyDeleted.push(delId);
-                        this.state.completedJobs.push(input);
-                        this._createFlash(input[1] + " has completed!");
-                        this.complete = true;
-                    }
-                }
-
-                else if (input.length == 4) { // add new jobs to the end of the list
-                    var inArr = false;
-                    for (var i = 0; i < this.state.currentJobs.length; i++) {
-                        if (input[0] == this.state.currentJobs[i][0]) {
-                            if(input[2] === input[3]){
-                                complete = this.state.currentJobs.splice(i, 1);  complete = complete[0];
-                                this.state.completedJobs.push(complete);
-                                index = this.state.selectedIDs.indexOf(complete[0]);
-                                delId = this.state.selectedIDs.splice(index, 1); // stop querying this job
-                                this.recentlyDeleted.push(delId);
-                                this.state.fullRefresh = 5;
-                                inArr = true;
-                                this._createFlash(input[1] + " has completed!");
-                                this.complete = true;
-                                break;
-                            }
-
-                            inArr = true;
-                            newJobs.push(this.state.currentJobs.splice(i, 1, input));
-                            break;
-                        }
-                    }
-                    if (!inArr){
-                        if(input[2] === input[3]){
-                                this.state.completedJobs.push(input);
-                                index = this.state.selectedIDs.indexOf(input[0]);
-                                delId = this.state.selectedIDs.splice(index, 1); // stop querying this job
-                                this.recentlyDeleted.push(delId);
-                                this._createFlash(input[1] + " has completed!");
-                                this.complete = true;
-                        }else {
-                            this.state.currentJobs.push(input);
-                        }
-
-
-                        this.state.fullRefresh = 5;
-                    }
-                }
-            }
-
-
-            // remove duplicates
-
-            if(this.state.completedJobs.length) {
-                var temp = new Set(this.state.completedJobs.map(i=> {
-                    return i[0];
-                }));
-
-                this.state.completedJobs = this.state.completedJobs.filter(job => {
-                    if(temp.has(job[0])){
-                        temp.delete(job[0]);
-                        return job;
-                    }
-                });
-            }
-            this.refresh();
-        },
-
-        /*
-        Take this.state.selectedIDs, merge them together into a query, and then make an xhrRequest if we don't already
-        have a request going on.
-        */
-        _xhrLoop: function () {
-            this.inLoop = true;
-            var queryIds = this.recentlyDeleted.length === 0 ? this.state.selectedIDs.filter(function (id) {
-                if (this.recentlyDeleted.length && this.recentlyDeleted.indexOf(id) === -1) {
-                    return id;
-                }
-                else if (!this.recentlyDeleted.length)
-                    return id;
-            }.bind(this)) : this.state.selectedIDs;
-            queryIds = new Set(queryIds);// remove duplicates
-            queryIds = Array.from(queryIds);
-            this.state.pollType = (typeof queryIds[0] == "number") ? "shapes" : "voxels";
-            var pollquery = this.state.pollType + "?" + this.state.pollType + "=" + queryIds.join("$$"); //FOR VOXELS, FORMAT IS "hash$$id1$$etc." while for SHAPES, it's "id1$$id2$$etc."
-            if (this.ajaxInProgress && !this.request) {
-                this.request = this._createXhrRequest(pollquery);
-                $(window).on("beforeunload", function (event) {// abort the call on page unload
-                    if (this.request && this.request.readyState !== XMLHttpRequest.DONE)
-                        this.request.abort();
-                    this.ajaxInProgress = false;
-                });
-
-                this.request.send();
-                this.inLoop = false;
-            }
-            else if(this.ajaxInProgress && this.request) {
-                return;
-            }
-            else {
-                if (this.request)
-                    this.request.abort();
-
-                var timeout = 5000;
-                this.refresh();
-                // this.pollTimeout = setTimeout(function () { this._xhrLoop() }.bind(this), timeout);
-            }
-        },
-
-        //xhrpolling body method, created the request and handlers
-        _createXhrRequest: function (pollquery) {
-            var timeout = this.state.pollingState === 0 ? 5000 : 100;
-            var xhr = new XMLHttpRequest();
-            xhr.open('GET', '/update/' + pollquery, true);// rework poll functino to be a parameter with ids for both voxels and shapes
-            xhr.onreadystatechange = function () {
-                if (xhr.readyState === XMLHttpRequest.DONE ) {
-                    if( xhr.status == 200) {
-                        var progress;
-                        try {
-                            progress = JSON.parse(xhr.response);
-                            this._updateJobs(progress.progress);
-                            if (this.pageReload) {
-                                setTimeout(function() {
-                                    this.request.abort();
-                                    if (this.complete)
-                                        window.location.href = window.location.href.substring(0, window.location.href.lastIndexOf('/'));
-                                }.bind(this), 1000);
-                                return;
-                            }
-                        } catch (e) {
-                            progress = JSON.parse(xhr.response);
-                            this._updateJobs(progress.progress);
-                            if (e.name !== "SyntaxError") {
-                                console.log('Server Error: Reponse cannot be parsed.' + e.toString());
-                                this._createFlash('Server Error - if this persists please contact the developers')
-                            }
-                        }
-                    }
-                    if (this.ajaxInProgress) {
-                        var timeout = this.state.pollingState === 0 ? 5000 : 100;
-                        this.pollTimeout = setTimeout(function () { this.request = null;this._xhrLoop();}.bind(this), timeout + 200);
-                    }
-                }
-            }.bind(this);
-            // xhr.timeout = timeout + 200;
-            xhr.responsetype = "json"
-            // xhr.ontimeout = function () {
-            //     if(this.state.timeouts > 0)
-            //          this.state.timeouts -= 1;
-
-            //     if (this.state.timeouts == 0) {
-            //         this.pollingState = 0;
-            //     }
-            //     var timeout = this.state.pollingState === 0 ? 5000 : 100;
-
-            //     if (!this.inLoop && this.request === null)
-            //     {
-            //         this.pollTimeout = setTimeout(function () { this._xhrLoop() }.bind(this), timeout +200);
-
-            //     }
-            // }.bind(this);
-            return xhr;
-        },
-
-        _defaultState: function () {
-            return {
-                currentJobs: [],
-                restCounter: 0,
-                pollType: "shapes",
-                selectedIDs: [],
-                timeouts: 5,
-                isDisplayed: false,
-                fullRefresh: 5,
-                completedJobs: []
-            }
-        }
-    });
-
-
-}
-
-
+///////////////////////////////////////////
+// The main entry point of progressWidget.
+///////////////////////////////////////////
 $(document).ready(() => {
     if (!$('.progress-tracker').length) {// only launch if a user is signed in
         return
@@ -701,10 +11,433 @@ $(document).ready(() => {
         .progressWidget();
 });
 
+////////////////////////////////////////////
+// The widget.
+////////////////////////////////////////////
+function progressWidgetInit(){
+    $.widget("custom.progressWidget", {
+        /**
+        * Widget constructor.
+        * @constructor
+        */
+        _create: function(){
+            this._loadVariables();
+            this._checkUrl();
+            // We access the code below if and only if we don't have anything new to query.
+            this._createWidgetMenu();
+            this._display();
+            // This is the main loop: we query the controller, update the data structures accordingly, and then
+            // update the widget to that effect.
+            setInterval(function(){
+                this._queryController();
+            }.bind(this), 1000);
+        },
+
+        /**
+        * Called after we've created the widget and binded actions. This will display if isDisplayed is
+        * true.
+        */
+        _display: function(){
+            if (this.state.isDisplayed)
+                this.$el.show();
+            else
+                this.$el.hide();
+        },
+
+        /**
+        * Called from _create; loads the local variables, and, if savedState is defined, that, too.
+        */
+        _loadVariables: function(){
+            // Loads the elements into more convenient names.
+            this.element;
+            this.$el = $(this.element);
+            this.$anchor = $('.progress-tracker')
+            this.$flashhandler = $('#flashes');
+            // Other stuff to keep track of.
+            this.jobsAdded = false; // If true, we only access _updateWidgetMenu. If false, we'll have to access _createWidgetMenu to update.
+            this.flashes = []; // The array of flashes.
+            this.reset = false; // Do we reset upon refresh?
+            // Stuff pertaining to this.state.queryIds.
+            this.request = null; // The current XMLHTTPRequest, when querying for job progress.
+            this.progress = []; // This gets populated upon querying the controller for job progress. Of course, the
+                                // hard part is comparing this to the current jobs (this.state.jobs)
+            // Loads the saved state.
+            var savedState = window.localStorage.getItem('progressWidget');
+            if (savedState && savedState !== 'undefined')
+                this.state = JSON.parse(savedState);
+            else
+                this.state = this._defaultState();
+        },
+
+        /**
+        * This processes the URL to see if any new uploads have occurred; if so, this is added to queryIds, and then we save our
+        * state to localStorage in preparation for a refresh.
+        */
+        _checkUrl: function(){
+            // Are we uploading layers? /layers/<user_id>/<file_id>
+            // Are we uploading voxels? /voxels/<user_id>/<hash_voxel_id>$$<...data_layer_id's> (joined by double dollars)
+            // What if we're not uploading anything? (layers) /layers/<user_id> OR (voxels) /voxels/<user_id>
+            var urlString = window.location.pathname;
+            var isLayerUrl = contains(urlString, '/layers'),
+                isVoxelUrl = contains(urlString, '/voxels'),
+                urlHasUploads = (urlString.match(/\/\d+(?!\.)\/\d+(?!\.)/) != null); // This matches two numbers with no decimal points.
+            if ((isLayerUrl || isVoxelUrl) && urlHasUploads) {
+                try {
+                    // We now split, and parse the URL. It differs if we have layers vs. voxels.
+                    var temp = urlString.split('/');
+                    var uploadInfo = temp[temp.length - 1];
+                    if (isLayerUrl){
+                        // Add layerInfo to the set of shapes to query.
+                        var layerInfo = parseInt(uploadInfo);
+                        this.state.queryIds.shapes.push(layerInfo);
+                    }
+                    else {
+                        // Slightly more complicated. Add {hash: {<id's>}} to the voxels to query.
+                        var voxelInfo = uploadInfo.split("$$");
+                        var hash = voxelInfo[0],
+                            dataLayerIds = voxelInfo.slice(1);
+                        this.state.queryIds.voxels[hash] = dataLayerIds;
+                    }
+                    // Save the state to localStorage; refresh the page.
+                    window.localStorage.setItem('progressWidget', JSON.stringify(this.state));
+                    window.location.href = window.location.href.substring(0, window.location.href.lastIndexOf('/'));
+                }
+                catch (e) {
+                    console.log('PROGRESSWIDGET ERROR: error adding ID, url: ' + urlString)
+                    console.log(this.uploadId)
+                }
+            }
+        },
+
+        /**
+        * Creates the base widget menu, when we're free to (after any ID's have been uploaded).
+        */
+        _createWidgetMenu: function(){
+            var $progressWidget = this.element,
+                widgetCaption = "This tracks all current jobs in the worker queue. hover over the job names to see how many shapes/voxels have been processed.",
+                refreshCaption = "Click here to reset the Progress Tracker.",
+                htmlBuilder = '';
+
+            htmlBuilder += `<div id="widget-menu">
+                <span class="tooltiper caption-left" data-tooltip="${widgetCaption}"><i class="fa fa-question" style="top:'0px!important'"/></span>
+                <span class="tooltiper caption-right" data-tooltip="${refreshCaption}"><i class="fa fa-refresh refresh-widget" style="top:'0px!important'"/></span>
+                <ul id="jobList"></ul>
+            </div>`;
+            this.state.jobs;
+            widgetMenu = $.parseHTML(htmlBuilder);
+            widgetMenu = $(widgetMenu);
+            widgetMenu.text = widgetCaption;
+            // Adds the job display to the widgetMenu, if there are any.
+            if (this.state.jobs.length > 0) {
+                this.state.jobs.forEach(job => {
+                    htmlBuilder = createProgressBar(...job);
+                    var temp = $.parseHTML(htmlBuilder);
+                    widgetMenu.find('#jobList').append(temp);
+                });
+            }
+            else {
+                widgetMenu.find('#jobList').append($.parseHTML(noJobs()));
+            }
+            // Now we remove the existing widget, and replace it with the widget we've just created.
+            if ($('#widget-menu').length)
+                $('#widget-menu').remove();
+
+            this.$widgetMenu = widgetMenu;
+            this.$el.append(this.$widgetMenu);
+            // toolTiper, do your magic!
+            toolTiper();
+            // Bind UI Actions.
+            this._bindUIActions();
+        },
+
+        /**
+        * Creates a flash.
+        */
+        _createFlash: function (input) {
+            this.$anchor.trigger('pflash', ["\t Progress Tracker: ", input]);
+        },
+
+        /**
+        * Binds UI actions to the widget.
+        */
+        _bindUIActions: function(){
+            // Event handler for closing out a job.
+            if (this.$widgetMenu.children().length) {
+                $('.close-btn').on('click', this._closeItem.bind(this));
+            }
+            // Event handler for resetting the widget, upon refresh.
+            this._on($('.refresh-widget'), {
+                click: '_toggleReset'
+            })
+            // Event handler for displaying the widget.
+            this._on(this.$anchor, {
+                click: '_toggleDisplay'
+            });
+            // Event listener for unloading a page.
+            window.addEventListener('beforeunload', this._destroy.bind(this), false);
+            // Event handler for whenever _createFlash is called.
+            this._on(this.$anchor, {
+                'pflash': function (e) {
+                    e.preventDefault();
+                    if (!e.handled){
+                        e.handled = true;
+                        var event = arguments[1],
+                            message = arguments[2];
+
+                        var flash = $('<div class="pflash alert alert-info alert-dismissible"> <strong>' + event + '</strong>' + message + '</div>');
+
+                        flash.on('click', function () {
+                            $(this).fadeOut(1000, function() {
+                                $(this).remove();
+                            });
+                        });
+                        var t = setTimeout(function () {
+                            $(flash).fadeOut(1000, function () {
+                                $(this).remove();
+                            });
+                            this.flashes.shift();
+                        }.bind(this).bind(flash), 5000); // This is the flashTimeout.
+                        this.flashes.push(t);
+                        $('.navbar-fixed-top').append(flash);
+                    }
+                }
+            });
+        },
+
+        /**
+        * Helper method for _bindUIActions.
+        * When we unload a page, this function is called. The state is saved and all components are removed.
+        */
+        _destroy: function () {
+            // Save the state to localStorage if this.reset is False, and vice versa.
+            if(!this.reset)
+                window.localStorage.setItem('progressWidget', JSON.stringify(this.state));
+            else
+                window.localStorage.removeItem('progressWidget');
+            // Hide the widget.
+            if (this.state.isDisplayed)
+                this.$el.hide();
+            // Remove the widget menu.
+            if (this.$widgetMenu) {
+                this.$widgetMenu.remove();
+            }
+            // Clear timeouts for any flashes.
+            if (this.flashes.length) {
+                this.flashes.forEach(item => {
+                    clearTimeout(item);
+                });
+            }
+            // Remove all children of this.$el.
+            if (this.$el && this.$el.children().length)
+                this.$el.children().remove();
+        },
+
+        /**
+        * Helper method for _bindUIActions.
+        * When the "Progress Tracker" button is clicked, the display is toggled.
+        * TODO (minor): Add poll query timing features. For now, we update every second or so, no matter the toggle.
+        */
+        _toggleDisplay: function(e) {
+            e.preventDefault();
+            if (!e.handled){
+                e.handled = true;
+                this.state.isDisplayed = !this.state.isDisplayed;
+                if (this.state.isDisplayed)
+                    this.$el.slideDown(200);
+                else
+                    this.$el.slideUp(200);
+            }
+        },
+
+        /**
+        * Helper method for _bindUIActions.
+        * When the refresh button is clicked, this function is called.
+        */
+        _toggleReset: function(e) {
+            e.preventDefault();
+            if (!e.handled){
+                e.handled = true;
+                this.reset = !this.reset;
+                messages = ["State will no longer reset on page refresh.", "State will reset on page refresh. To cancel, click the refresh button again."];
+                this._createFlash(messages[+this.reset]);
+                return;
+            }
+        },
+
+        /**
+        * Helper method for _bindUIActions.
+        * When an "x" button is clicked, a job is closed out.
+        * TODO: decide whether or not the commented out code should be included.
+        */
+        _closeItem: function(e) {
+            e.preventDefault();
+            if (!e.handled){
+                e.handled = true;
+                // Physically remove the item from the UI.
+                var menuItem = $(e.target).closest('.menu-item');
+                var id = menuItem.attr('id');
+                menuItem.remove();
+                // Some error message
+                if (id === null) {
+                    console.log('UPLOAD WIDGET Error: False destroychild id' + id)
+                    this._createFlash('Error destroying child, reset local storage or contact developers if this persists.')
+                    return false;
+                }
+                // Manage the arrays.
+                for (var i = 0; i < this.state.jobs.length; i++) {
+                    if (this.state.jobs[i][0] == id){
+                        // Remove the job from this.state.queryIds.
+                        this._removeFromQueryIds(this.state.jobs[i]);
+                        // Remove the job from this.state.jobs.
+                        this.state.jobs.splice(i, 1);
+                    }
+                }
+                this._createWidgetMenu();
+            }
+        },
+
+        /**
+        * TODO: given a set of ID's to query (both voxels and shapes), query the controller for the status of said items.
+        */
+        _queryController: function(){
+            // This assembles all of the queries from queryIds to a string:
+            // [id1$$id2$$...][hash1$numIds1$$hash2$numIds2$$...]
+            var arrOfQueries = this.state.queryIds.shapes.slice(0);
+            $.each(this.state.queryIds.voxels, function(hash, dataLayerIds){
+                arrOfQueries.push(hash + '$' + dataLayerIds.length);
+            });
+            var queryParam = "shapes?shapes=" + arrOfQueries.join('$$');
+            // Now, we ask AJAX about this URL.
+            this.request = this._createXhrRequest(queryParam);
+            $(window).on("beforeunload", function (event) { // Abort the call on page unload.
+                if (this.request && this.request.readyState !== XMLHttpRequest.DONE)
+                    this.request.abort();
+            });
+            this.request.send();
+        },
+
+        /**
+        *
+        */
+        _createXhrRequest: function (pollquery) {
+            var xhr = new XMLHttpRequest();
+            xhr.open('GET', '/update/' + pollquery, true);
+            xhr.onreadystatechange = function () {
+                if (xhr.readyState === XMLHttpRequest.DONE && xhr.status == 200) {
+                    var progress = JSON.parse(xhr.response);
+                    this.progress = progress.progress;
+                    // We now update the arrays and the widget menu.
+                    if (this.progress && this.progress != "undefined") {
+                        this._updateArrays();
+                        if (!this.jobsAdded){
+                            this._createWidgetMenu();
+                            this.jobsAdded = true;
+                        }
+                        else {
+                            this._updateWidgetMenu();
+                        }
+                    }
+                }
+            }.bind(this);
+            xhr.responsetype = "json";
+            return xhr;
+        },
+
+        /**
+        * Given the results from _queryController, we update the arrays.
+        * TODO: update updateController.js to fit our parameters.
+        */
+        _updateArrays: function(){
+            // First, update this.state.jobs with info from this.progress.
+            // If there's a job in this.progress that's not in this.state.jobs, add it. There should be at most 1.
+            this.progress; // Reminder: this shows an array of 6-length arrays: [id, jobName, num, den, pollType, completed]
+            for (var i in this.progress){
+                var job = this.progress[i];
+                var newJob = true;
+                for (var j in this.state.jobs){
+                    var candJob = this.state.jobs[j];
+                    // Does the job ID match the job ID that we're considering?
+                    if (job[0] == candJob[0]){
+                        // If so, then update the numerator and denominator.
+                        candJob[2] = job[2];
+                        candJob[3] = job[3];
+                        newJob = false;
+                    }
+                }
+                // If not for anything in this.state.jobs, add it.
+                if (newJob){
+                    this.state.jobs.push(job);
+                }
+            }
+            // Next, check the jobs list to see if anything has completed. If so, remove the job ID from
+            // this.state.queryIds, and create a flash.
+            for (var i in this.state.jobs){
+                var job = this.state.jobs[i];
+                if (job[2] >= job[3] && !job[5]){
+                    job[5] = true;
+                    this._createFlash(job[1] + " has completed!");
+                    this._removeFromQueryIds(job);
+                }
+            }
+        },
+
+        /**
+        * Given a job (6-length array etc.), the job is removed from this.state.queryIds.
+        */
+        _removeFromQueryIds: function(job){
+            if (job[4] == "voxels"){ // JSON vs Set. Lol.
+                delete this.state.queryIds.voxels[job[0]];
+            }
+            if (job[4] == "shapes"){
+                var index = this.state.queryIds.shapes.indexOf(job[0]);
+                this.state.queryIds.shapes.splice(index, 1);
+            }
+        },
+
+        /**
+        * Updates the widget menu based on what happened to the arrays.
+        * TODO (minor): on job complete, the text should change!
+        */
+        _updateWidgetMenu: function(){
+            this.state.jobs.forEach(job => {
+                var [id, jobName, numerator, denominator, type, completed] = job,
+                    percentage = ((numerator / denominator) * 100).toFixed(2).toString(),
+                    innerString = jobName + ": " + numerator + "/" + denominator,
+                    frac = numerator + "/" + denominator,
+                    menuItem = $('#' + id),
+                    datatool = menuItem.find('.tooltiper').attr('data-tooltip', frac)
+                menuItem.find('.progress-label').text(percentage + "%");
+                menuItem.find('.progress-bar').css('width', percentage + "%");
+            });
+        },
+
+        /**
+        * Loads the default state.
+        */
+        _defaultState: function(){
+            return {
+                isDisplayed: false, // Default: don't display.
+                queryIds: {
+                    shapes: [], // An array of ID's.
+                    voxels: {},  // A JSON, mapping a voxel's hash property to an array of its corresponding data layer IDs.
+                },
+                jobs: [] // Each element of the array is of the form: [id, jobName, numerator, denominator, type (voxel or shape), completed (boolean)]
+            };
+        },
+    });
+}
+
+///////////////////////////////////////////
+// The rest are helper methods.
+///////////////////////////////////////////
+function contains(listable, match){
+    return listable.indexOf(match) != -1;
+}
+
 function clearHTML() { // helper method that returns cleared inner html section
     var progressList = document.getElementById('progressList');
     if (progressList && progressList.hasChildNodes()) {
-        progressList.innesrHTML = '';
+        progressList.innerHTML = '';
     }
     return progressList
 }
@@ -715,34 +448,12 @@ function noJobs() {// return div with default message for no currnet jobs
     </div>`
 }
 
-function createCompleteBar(id, jobName) { //clear a single progress bar DOM Element
-    var numerator = 100, denominator = 100;
-    jobname = escapeHTML(jobName);
-    var percentage = "100/100",
-        innerString = jobName + ": 100/100",
-        frac = "Job Complete";
-    var html = `
-    <div id="${id}" class="menu-item" style="">
-        <div class="container">
-        <span class="tooltiper" data-tooltip="${frac}"> ${jobName}</span>
-        <span class="progress-label" style=""> ${percentage} </span>
-        <div class="progress-bar progress-bar-success" role="progressbar" aria-valuenow="40"
-            aria-valuemin="3" aria-valuemax="100" style="min-width: 1%; width:`;
-    html += percentage + `%\">
-        </div>
-        <i id="${id}" class="fa fa-times close-btn" aria-hidden="true"></i>
-        </div>
-    </div>
-
-    `;
-    return html
-}
-
-function createProgressBar(id, jobName, numerator, denominator) { //clear a single progress bar DOM Element
+function createProgressBar(id, jobName, numerator, denominator, type) { //clear a single progress bar DOM Element
     jobname = escapeHTML(jobName);
     var percentage = ((numerator / denominator) * 100).toFixed(2).toString(),
         innerString = jobName + ": " + numerator + "/" + denominator,
-        frac = numerator + "/" + denominator + " processed";
+        // If complete, just say "Job Complete"
+        frac = (numerator == denominator) ? "Job Complete" : (numerator + "/" + denominator + " processed");
     var html = `
     <div id="${id}" class="menu-item" style="">
         <div class="container">
